Add files via upload
Scenario, given a set of current assignment grades, and assuming  that every future assignment you earn a 10 (max points). Find the minimum number of future assignments needed to get a final grade of 10, which you would receive if your average is 9.5 or above.

Example: current grades are 8 and 9, you would need a minimum of four assignments to get a final grade of 10.